move to viz context
